Move check_second_stage_environment to AutoinstFunctions
