+++ src/modules/AutoinstConfig.rb
@@ -35,8 +35,6 @@ module Yast
       Yast.import "Label"
       Yast.import "Report"
       Yast.import "Profile"
-      Yast.import "AutoinstFunctions"
-      Yast.import "Pkg"
 
       Yast.include self, "autoinstall/xml.rb"
 
@@ -229,38 +227,6 @@ module Yast
       @Proposals = deep_copy(l)
 
       nil
-    end
-
-    # Checking the environment the installed system
-    # to run a second stage if it is needed.
-    #
-    # @return [String] empty String or error messsage about missing packages.
-    def check_second_stage_environment
-      error = ""
-      return error unless AutoinstFunctions.second_stage_required?
-
-      missing_packages = Profile.needed_second_stage_packages.select do |p|
-        !Pkg.IsSelected(p)
-      end
-      unless missing_packages.empty?
-        log.warn "Second stage cannot be run due missing packages: #{missing_packages}"
-        # TRANSLATORS: %s will be replaced by a package list
-        error = format(_("AutoYaST cannot run second stage due to missing packages \n%s.\n"),
-          missing_packages.join(", "))
-        unless registered?
-          if Profile.current["suse_register"] &&
-            Profile.current["suse_register"]["do_registration"] == true
-            error << _("The registration has failed. " \
-              "Please check your registration settings in the AutoYaST configuration file.")
-            log.warn "Registration has been called but has failed."
-          else
-            error << _("You have not registered your system. " \
-              "Missing packages can be added by configuring the registration in the AutoYaST configuration file.")
-            log.warn "Registration is not configured at all."
-          end
-        end
-      end
-      error
     end
 
     # Searches for 'autoyast' via SLP and returns the full URL of
@@ -671,16 +637,6 @@ module Yast
     publish :function => :check_second_stage_environment, :type => "string ()"
 
     private
-
-    # Determine whether the system is registered
-    #
-    # @return [Boolean]
-    def registered?
-      require "registration/registration"
-      Registration::Registration.is_registered?
-    rescue LoadError
-      false
-    end
 
     # Reads base product name from the profile
     #
