+++ src/modules/AutoinstFunctions.rb
@@ -1,6 +1,8 @@
 module Yast
   # Helper methods to be used on autoinstallation.
   class AutoinstFunctionsClass < Module
+    include Yast::Logger
+
     def main
       textdomain "installation"
 
@@ -8,6 +10,7 @@ module Yast
       Yast.import "Mode"
       Yast.import "AutoinstConfig"
       Yast.import "ProductControl"
+      Yast.import "Pkg"
     end
 
     # Determines if the second stage should be executed
@@ -29,6 +32,51 @@ module Yast
         ProductControl.RunRequired("continue", Mode.mode)
       end
     end
+
+    # Checking the environment the installed system
+    # to run a second stage if it is needed.
+    #
+    # @return [String] empty String or error messsage about missing packages.
+    def check_second_stage_environment
+      error = ""
+      return error unless second_stage_required?
+
+      missing_packages = Profile.needed_second_stage_packages.select do |p|
+        !Pkg.IsSelected(p)
+      end
+      unless missing_packages.empty?
+        log.warn "Second stage cannot be run due missing packages: #{missing_packages}"
+        # TRANSLATORS: %s will be replaced by a package list
+        error = format(_("AutoYaST cannot run second stage due to missing packages \n%s.\n"),
+          missing_packages.join(", "))
+        unless registered?
+          if Profile.current["suse_register"] &&
+            Profile.current["suse_register"]["do_registration"] == true
+            error << _("The registration has failed. " \
+              "Please check your registration settings in the AutoYaST configuration file.")
+            log.warn "Registration has been called but has failed."
+          else
+            error << _("You have not registered your system. " \
+              "Missing packages can be added by configuring the registration in the AutoYaST configuration file.")
+            log.warn "Registration is not configured at all."
+          end
+        end
+      end
+      error
+    end
+
+  private
+
+    # Determine whether the system is registered
+    #
+    # @return [Boolean]
+    def registered?
+      require "registration/registration"
+      Registration::Registration.is_registered?
+    rescue LoadError
+      false
+    end
+
   end
 
   AutoinstFunctions = AutoinstFunctionsClass.new
