+++ test/AutoinstFunctions_test.rb
@@ -64,4 +64,48 @@ describe Yast::AutoinstFunctions do
       end
     end
   end
+
+  describe "#check_second_stage_environment" do
+    context "second stage is not needed" do
+      it "returns an empty error string" do
+        allow(subject).to receive(:second_stage_required?).and_return(false)
+        expect(subject.check_second_stage_environment).to be_empty
+      end
+    end
+
+    context "second stage is needed" do
+      before do
+        allow(subject).to receive(:second_stage_required?).and_return(true)
+      end
+
+      context "required package are installed" do
+        it "returns an empty error string" do
+          allow(Yast::Pkg).to receive(:IsSelected).and_return(true)
+          expect(subject.check_second_stage_environment).to be_empty
+        end
+      end
+
+      context "required package are not installed" do
+        before do
+          allow(Yast::Pkg).to receive(:IsSelected).and_return(false)
+        end
+
+        context "registration has not been defined in AY configuration file" do
+          it "reports error to set registration" do
+            allow(Yast::Profile).to receive(:current).and_return({})
+            expect(subject.check_second_stage_environment).to include("configuring the registration")
+          end
+        end
+
+        context "registration has failed" do
+          it "reports error to check registration settings" do
+            allow(Yast::Profile).to receive(:current).and_return(
+              {"suse_register" => {"do_registration" => true}})
+            expect(subject.check_second_stage_environment).to include("registration has failed")
+          end
+        end
+      end
+    end
+  end
+
 end
